Time: 3 ms (92.00%), Space: 6.3 MB (89.44%) - LeetHub

+++ 1220-count-vowels-permutation/1220-count-vowels-permutation.cpp
@@ -12,6 +12,6 @@ public:
             aCnt = aNewCnt, eCnt = eNewCnt, iCnt = iNewCnt, oCnt = oNewCnt, uCnt = uNewCnt;
         }
         
-        return (aCnt + eCnt + iCnt + oCnt + uCnt) % MOD;
+        return static_cast<int>((aCnt + eCnt + iCnt + oCnt + uCnt) % MOD);
     }
 };
